pyproject.toml
[tool.poe]
envfile = ".env"

[tool.poe.tasks]
help.cmd = "poe tasks"
help.help = "Show help"

lint.cmd = "bash ./scripts/lint.sh"
lint.help = "Run linter"

format.cmd = "bash ./scripts/format.sh"
format.help = "Run formatter"

test.cmd = "bash ./scripts/test.sh"
test.help = "Run tests"

super-user.shell = """docker exec excuseme-backend sh -c "uv run python manage.py createsuperuser" """
super-user.help = "Create super user"

make-migrations.shell = """docker exec excuseme-backend sh -c "uv run python manage.py makemigrations" """
make-migrations.help = "Make migrations"

migrate.shell = """docker exec excuseme-backend sh -c "uv run python manage.py migrate" """
migrate.help = "Migrate"

build-dev.cmd = "docker compose -f docker-compose.yml up --build -d"
build-dev.help = "Build and run dev environment"

stop-dev.cmd = "docker compose -f docker-compose.yml down"
stop-dev.help = "Stop dev environment"

build-prod.cmd = "docker compose -f docker-compose.prod.yml up --build -d"
build-prod.help = "Build and run prod environment"

stop-prod.cmd = "docker compose -f docker-compose.prod.yml down"
stop-prod.help = "Stop prod environment"

reset.shell = """docker system prune -a --volumes -f"""
reset.help = "Clean up all docker resources"

list.shell = """docker images && docker ps -a"""
list.help = "List docker images and containers"

docker-logs.shell = """docker logs excuseme-backend && docker logs excuseme-celery-worker && docker logs excuseme-celery-beat"""
docker-logs.help = "List docker logs of backend, celery-worker and celery-beat"




[project]
name = "excuseme"
version = "0.1.0"
authors = ["Atlas-Zero", "Fabian993"]
dependencies = [
    "amqp==5.1.1",
    "asgiref==3.6.0",
    "billiard==3.6.4.0",
    "black==23.3.0",
    "celery==5.2.7",
    "click==8.1.3",
    "click-didyoumean==0.3.0",
    "click-plugins==1.1.1",
    "click-repl==0.2.0",
    "coverage[toml]==7.2.5",
    "django==4.2.1",
    "django-cors-headers==3.14.0",
    "djangorestframework>=3.16.1",
    "gunicorn==20.1.0",
    "kombu==5.2.4",
    "poethepoet>=0.38.0",
    "prompt-toolkit==3.0.38",
    "psycopg==3.1.9",
    "psycopg2-binary>=2.9.11",
    "pytest==7.3.1",
    "pytest-django==4.5.2",
    "pytest-sugar==0.9.7",
    "pytz==2023.3",
    "redis==4.5.5",
    "ruff==0.0.265",
    "six==1.16.0",
    "sqlparse==0.4.4",
    "typing-extensions==4.5.0",
    "vine==5.0.0",
    "wcwidth==0.2.6",
    "whitenoise==6.4.0",
]


# TESTING
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --force-sugar --no-migrations --reuse-db --log-cli-level=INFO"
testpaths = [
    "tests",
]
pythonpath = [".", "src"]
python_files = "tests.py test_*.py *_tests.py"
DJANGO_SETTINGS_MODULE = "excuseme.settings.test"
filterwarnings = [
    'ignore::DeprecationWarning:kombu.*:',
    'ignore::DeprecationWarning:celery.*:',
]

[tool.coverage.report]
fail_under = 85
show_missing = "true"
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "import*",
    "def __str__",
    "def on_success",
    "def clean",
    "if missing",
    "if relations.exists()",
    "(FileDoesNotExistException, FileNotSupportedException)",
]


[tool.coverage.run]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/urls.py",
    "*/settings/*",
    "*/wsgi.py",
    "manage.py",
    "*__init__.py",
]
source = ["src"]


# LINTING
[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
# 'extend-exclude' excludes files or directories in addition to the defaults
extend-exclude = '''
^(.*/)?migrations/.*$
'''


[tool.ruff]
format = "grouped"
line-length = 88 # black default
extend-exclude = [
    "src/migrations/*",
    "src/media/*",
    "src/static/*",
    "src/manage.py",
    "*/test_data/*",
    "*__init__.py",
]

select = ["E", "F"]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "F405", # name may be undefined, or defined from star imports
]


# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.11.
target-version = "py311"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10


[tool.ruff.isort]
force-to-top = ["src"]
known-first-party = ["src"]
